requirements.txt
streamlit
pandas
openai
Pillow
librosa
matplotlib
numpy
openai
matplotlib
streamlit
